rename MenuRight -> RightMenu
