+++ superset-frontend/src/views/components/Menu.tsx
@@ -28,7 +28,7 @@ import { Link } from 'react-router-dom';
 import Icons from 'src/components/Icons';
 import { useUiConfig } from 'src/components/UiConfigContext';
 import { URL_PARAMS } from 'src/constants';
-import RightMenu from './MenuRight';
+import RightMenu from './RightMenu';
 import { Languages } from './LanguagePicker';
 
 interface BrandProps {
