+++ superset-frontend/src/views/components/RightMenu.tsx
